# swagger.yml
swagger: "2.0"
info:
  description: "Trustbrix - real estate for everyone"
  version: "1.0.0"
  title: "Trustbrix"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "developers@trustbrix.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:4500"
basePath: "/api"
...
schemes:
- "https"
- "http"
paths:
...
  /forget-password/{id}:
    post:
      tags:
      - "Users"
      summary: "Get user by id"
      description: ""
      operationId: "forgetPassword"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The id that needs to be fetched. "
        required: true
        type: "string"
      - name: "body"
        in: "body"
        required: true
        schema:
          $ref: "#/definitions/forget"
      responses:
        200:
          description: "successful operation"
        400:
          description: "Invalid username supplied"
        404:
          description: "User not found"
definitions:
...
  forget:
    type: "object"
    required:
    - "password"
    properties:
      password:
        type: "string"
        example: "test@123"
...
securityDefinitions:
  UserSecurity:
    type: apiKey
    in: header
    name: access-token
security:
  - UserSecurity: []                
...